+++ metastore/metastore-api/src/main/java/org/apache/drill/metastore/statistics/ColumnStatistics.java
@@ -28,7 +28,6 @@ import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.datatype.joda.JodaModule;
 import org.apache.drill.common.types.TypeProtos;
-import org.apache.drill.common.util.JacksonUtils;
 import org.apache.drill.metastore.util.TableMetadataUtils;
 
 import java.io.IOException;
@@ -67,9 +66,10 @@ import java.util.stream.Collectors;
 @JsonPropertyOrder({"statistics", "comparator"})
 public class ColumnStatistics<T> {
 
-  private static final ObjectMapper MAPPER = JacksonUtils.createJsonMapperBuilder()
-      .addModule(new JodaModule())
-      .build();
+  private static final ObjectMapper MAPPER =
+      StatisticsHolder.createJsonMapperBuilder()
+        .addModule(new JodaModule())
+        .build();
 
   private static final ObjectWriter OBJECT_WRITER = MAPPER.writerFor(ColumnStatistics.class);
